services/payment_service/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
debugpy
yookassa
sentry-sdk[fastapi]